+++ internal/crypto/tokens.go
@@ -1,9 +1,11 @@
 package crypto
 
 import (
+	"crypto/sha256"
 	"errors"
 	"fmt"
 
+	"github.com/nspcc-dev/neo-go/pkg/crypto/hash"
 	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 )
@@ -11,11 +13,15 @@ import (
 // TODO: https://github.com/nspcc-dev/neofs-node/issues/2795 after API stabilization, move some components to SDK
 
 // AuthenticateToken checks whether t is signed correctly by its issuer.
+//
+// If signature scheme is unsupported, [ErrUnsupportedScheme] returns. It also
+// returns when [neofscrypto.N3] scheme is used but fsChain is not provided.
 func AuthenticateToken[T interface {
 	SignedData() []byte
 	Signature() (neofscrypto.Signature, bool)
 	Issuer() user.ID
-}](token T) error {
+	Iat() uint64
+}](token T, fsChain HistoricN3ScriptRunner) error {
 	issuer := token.Issuer()
 	if issuer.IsZero() {
 		return errors.New("missing issuer")
@@ -24,9 +30,10 @@ func AuthenticateToken[T interface {
 	if !ok {
 		return errMissingSignature
 	}
+	var signer user.ID
 	switch scheme := sig.Scheme(); scheme {
 	default:
-		return fmt.Errorf("unsupported scheme %v", scheme)
+		return ErrUnsupportedScheme(scheme)
 	case neofscrypto.ECDSA_SHA512, neofscrypto.ECDSA_DETERMINISTIC_SHA256, neofscrypto.ECDSA_WALLETCONNECT:
 		pub, err := decodeECDSAPublicKey(sig.PublicKeyBytes())
 		if err != nil {
@@ -35,9 +42,22 @@ func AuthenticateToken[T interface {
 		if !verifyECDSAFns[scheme](*pub, sig.Value(), token.SignedData()) {
 			return schemeError(scheme, errSignatureMismatch)
 		}
-		if user.NewFromECDSAPublicKey(*pub) != issuer {
-			return errIssuerMismatch
+		signer = user.NewFromECDSAPublicKey(*pub)
+	case neofscrypto.N3:
+		if fsChain == nil {
+			return ErrUnsupportedScheme(neofscrypto.N3)
 		}
+		verifScript := sig.PublicKeyBytes()
+		verifScriptHash := hash.Hash160(verifScript) // TODO: or cut issuer?
+		if err := verifyN3ScriptsAtEpoch(fsChain, token.Iat(), verifScriptHash, sig.Value(), verifScript, func() [sha256.Size]byte {
+			return sha256.Sum256(token.SignedData())
+		}); err != nil {
+			return err
+		}
+		signer = user.NewFromScriptHash(verifScriptHash) // TODO: or check signer before?
+	}
+	if signer != issuer {
+		return errIssuerMismatch
 	}
 	return nil
 }